Fix a temporary bound warning
The vertex index property map of a boost CSR graph is stored inside
the counting_table object to be accessed whenever this object needs
to map a vertex to an index.
As described in the property map documentation:
"Copying a property map should be inexpensive since they are often
passed by value."
So simply keep a real copy instead of a reference.